[TS] modify the type of segment index from long to int

+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/NetworkSequenceViewReader.java
@@ -67,7 +67,7 @@ public interface NetworkSequenceViewReader {
     ResultSubpartitionView.AvailabilityWithBacklog getAvailabilityAndBacklog();
 
     /** Notify the required segmentId. */
-    void notifyRequiredSegmentId(long segmentId);
+    void notifyRequiredSegmentId(int segmentId);
 
     boolean isRegisteredAsAvailable();
