+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/consumer/RemoteInputChannel.java
@@ -828,7 +828,7 @@ public class RemoteInputChannel extends InputChannel {
     }
 
     @Override
-    public void notifyRequiredSegmentId(long segmentId) {
+    public void notifyRequiredSegmentId(int segmentId) {
         partitionRequestClient.notifyRequiredSegmentId(segmentId, this);
     }
